# docs/api/apex-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LogEntryOut": {
        "properties": {
          "at": {
            "title": "At",
            "type": "string"
          },
          "fields": {
            "additionalProperties": true,
            "description": "Provider extras not consumed by the mapped columns.",
            "title": "Fields",
            "type": "object"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "service": {
            "title": "Service",
            "type": "string"
          }
        },
        "required": [
          "at",
          "level",
          "service",
          "message"
        ],
        "title": "LogEntryOut",
        "type": "object"
      },
      "LogQueryIn": {
        "additionalProperties": false,
        "properties": {
          "filters": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "ANDed exact-match filters (e.g. service, level); 'thread_id' deep-links a pipeline run's logs by convention.",
            "title": "Filters",
            "type": "object"
          },
          "text": {
            "anyOf": [
              {
                "maxLength": 2048,
                "pattern": "^[^\\x00]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Free-text query (Lucene query_string syntax on ELK).",
            "title": "Text"
          }
        },
        "title": "LogQueryIn",
        "type": "object"
      },
      "LogSearchRequest": {
        "additionalProperties": false,
        "properties": {
          "connection_id": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 1,
                "pattern": "^[^\\x00]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Id"
          },
          "limit": {
            "default": 50,
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "default": 0,
            "exclusiveMaximum": 10000.0,
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "query": {
            "$ref": "#/components/schemas/LogQueryIn"
          },
          "window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WindowIn"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to the last hour, computed server-side."
          }
        },
        "title": "LogSearchRequest",
        "type": "object"
      },
      "LogSearchResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LogEntryOut"
            },
            "title": "Entries",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "window": {
            "$ref": "#/components/schemas/WindowOut"
          }
        },
        "required": [
          "entries",
          "total",
          "limit",
          "offset",
          "window"
        ],
        "title": "LogSearchResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WindowIn": {
        "additionalProperties": false,
        "properties": {
          "from": {
            "anyOf": [
              {
                "maxLength": 64,
                "pattern": "^[^\\x00]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO-8601 lower bound.",
            "title": "From"
          },
          "to": {
            "anyOf": [
              {
                "maxLength": 64,
                "pattern": "^[^\\x00]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO-8601 upper bound.",
            "title": "To"
          }
        },
        "title": "WindowIn",
        "type": "object"
      },
      "WindowOut": {
        "properties": {
          "from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From"
          },
          "to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To"
          }
        },
        "title": "WindowOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "APEX Orchestration Engine \u2014 Domain API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/logs/search": {
      "post": {
        "description": "Search logs through the configured LOG_SEARCH connection (any authenticated role).",
        "operationId": "searchLogs",
        "parameters": [
          {
            "description": "Explicit LOG_SEARCH connection to search through (overrides body.connection_id).",
            "in": "query",
            "name": "connection_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "minLength": 1,
                  "pattern": "^[^\\x00]*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Explicit LOG_SEARCH connection to search through (overrides body.connection_id).",
              "title": "Connection Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Logs",
        "tags": [
          "logs"
        ]
      }
    },
...
  }
}